Support variable length data records (#1)
* initial implementation to support variable length data

* [tests] Update SLOT_HEADER_SIZE and object size to be 8 bytes so it reflects the new changes to ringfs

* [test fix] ringfs_scan

Adjust calculations of slots per sector to match the altered slot scheme

* rename test.c -> test_ringfs.c

For my own sake, really

* [test_ringfs_append] Update asserts to match new slot scheme

* [test_ringfs_append] Let 'obj' be of type object_t so that sizes make sense, ffs

* [test_ringfs_discard] Update the test assertions to account for how variable length records behave

The original assertions have detailed knownledge about locations in the
fs. This is not so easy to do when records can be of variable length.
This is compensated in some cases and some assertions had to be removed.

* [ringfs] Fix some quirks of the scan logic

* [test_ringfs_capacity] Update assertion

For this scheme of variable length records, ringfs divides sectors into
'slots' and do not directly distinguish between the header and the data.
So the capacity is in terms of slots instead of number of objects as
before

* [ringfs] Properly implement ringfs_count_exact with new slot scheme

* [test_ringfs_count] Pass after some modification, see below

Had to increase size of the filesystem since it wrapped around with the
original setup: the test tried to add 10 records but was only able to
hold 5, so doubled the sector_size.

The count estimation does not work with the new slot scheme, so decided
to remove these assertions. I do not think it's possible to reliably
estimate the count with this approach to variable length records

* [test_ringfs_overflow] Modify to work with new slot scheme

* [test] Update description of the default filesystem used for testing

* [ringfs] Add APIs for appending and fetching variable length data records

* Encode the data type and length in the slot header
* The new APIs accepts a type parameter
* Length is a pointer when appending and updated on successful return

* [test] Append and fetch variable length data records

* Clean up by removing commented-out code